migrate assist to ChatCompletion API
* Break %%assist up into context gathering, generation, and magic
* Use the new Chat Completion API from OpenAI
