+++ genai/components.py
@@ -1,3 +1,4 @@
+import random
 import time
 
 from vdom import b as bold
@@ -58,3 +59,40 @@ def collapsible_log(children=None, title=None):
         title = [bold("Debug Log"), " - ", italics("click to expand/collapse")]
 
     return details(summary(title), div(children))
+
+
+def starting_message():
+    return random.choice(
+        [
+            "Phoning a friend 📲",
+            "Reaching out to another data scientist 📊",
+            "Just a little bit of data engineering will fix this 🔧",
+            "Trying my best 💯",
+            "Generating some code cells 💻",
+            "Asking the internet 🌐",
+            "Searching through my memory 💾",
+            "What would a data analyst do? 🤔",
+            "Querying my database 🗃️",
+            "Running some tests 🏃‍",
+            "One code cell, coming right up! 🚀",
+            "I'm a machine, but I still enjoy helping you code. 😊",
+        ]
+    )
+
+
+def completion_made():
+    return random.choice(
+        [
+            "Enjoy your BRAND NEW CELL 🚙",
+            "Just what you needed - more code cells! 🙌",
+            "Here's to helping you code! 💻",
+            "Ready, set, code! 🏁",
+            "Coding, coding, coding... 🎵",
+            "Just another code cell... 🙄",
+            "Here's a code cell to help you with your analysis! 📊",
+            "Need a code cell for your data engineering work? I got you covered! 🔥",
+            "And now for something completely different - a code cell! 😜",
+            "I got a little creative with this one - hope you like it! 🎨",
+            "This one's for all the data nerds out there! 💙",
+        ]
+    )
